Move the definition of LightningMessage class to lib/lightning/wire/lightning_messages/lightning_message.rb
