+++ lib/lightning/wire/lightning_messages.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'lightning/wire/lightning_messages/lightning_message.pb'
-
 module Lightning
   module Wire
     module LightningMessages
@@ -21,46 +19,7 @@ module Lightning
       module RoutingMessage
       end
 
-      module GossipQuery
-      end
-
-      class LightningMessage < Lightning::Wire::LightningMessages::Generated::LightningMessage
-        def self.load(payload)
-          type = payload.unpack('n').first
-          clazz = [
-            Init,
-            Error,
-            Ping,
-            Pong,
-            OpenChannel,
-            AcceptChannel,
-            FundingCreated,
-            FundingSigned,
-            FundingLocked,
-            Shutdown,
-            ClosingSigned,
-            UpdateAddHtlc,
-            UpdateFulfillHtlc,
-            UpdateFailHtlc,
-            UpdateFailMalformedHtlc,
-            CommitmentSigned,
-            RevokeAndAck,
-            UpdateFee,
-            ChannelReestablish,
-            AnnouncementSignatures,
-            ChannelAnnouncement,
-            NodeAnnouncement,
-            ChannelUpdate,
-            GossipTimestampFilter,
-            QueryChannelRange,
-            QueryShortChannelIds,
-            ReplyChannelRange,
-            ReplyShortChannelIdsEnd
-          ].select{|clazz| clazz.const_get('TYPE') == type}.first
-          return nil unless clazz
-          clazz.load(payload)
-        end
-      end
+      autoload :LightningMessage, 'lightning/wire/lightning_messages/lightning_message'
 
       autoload :Init, 'lightning/wire/lightning_messages/init'
       autoload :Error, 'lightning/wire/lightning_messages/error'
